Move implementation of timeTracker
